Fixed warnings in Back end

+++ modules/core/src/com/company/platform/core/FFMpegCaptureStream.java
@@ -1,10 +1,8 @@
 package com.company.platform.core;
 
 import com.company.platform.entity.Camera;
-import com.haulmont.cuba.core.sys.AppContext;
 import org.apache.commons.io.filefilter.WildcardFileFilter;
 import org.bytedeco.javacv.FrameGrabber;
-import org.bytedeco.javacv.FrameRecorder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.annotation.Scope;
@@ -14,9 +12,6 @@ import java.io.File;
 import java.io.FileFilter;
 import java.io.IOException;
 import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.Arrays;
 
 @Component(FFMpegCaptureStream.NAME)
 @Scope("prototype")
@@ -31,15 +26,12 @@ public class FFMpegCaptureStream extends AbstractFFMpegCapture {
 
     @Override
     protected void createFile() {
-        StringBuilder pathBuilder = new StringBuilder();
-        pathBuilder.append("file/")
-                .append(camera.getName())
-                .append(".m3u8");
 
-        String path = pathBuilder.toString();
+        String path = "file/" +
+                camera.getName() +
+                ".m3u8";
 
-        File file = new File(path);
-        this.file = file;
+        this.file = new File(path);
         log.info("File has been created");
     }
 
@@ -62,23 +54,24 @@ public class FFMpegCaptureStream extends AbstractFFMpegCapture {
     public void stop(){
         super.stop();
 
-        final String dilimeter = "/";
-        String[] parts = this.file.getPath().split(dilimeter);
+        final String delimiter = "/";
+        String[] parts = this.file.getPath().split(delimiter);
         StringBuilder tsFileBuilder = new StringBuilder();
         for (int i = 0; i < parts.length - 1; i++) {
             tsFileBuilder.append(parts[i])
-                    .append(dilimeter);
+                    .append(delimiter);
         }
         File tsFile = new File(tsFileBuilder.toString());
         FileFilter filter = new WildcardFileFilter(camera.getName() + "*.ts");
         File[] files = tsFile.listFiles(filter);
-        for(File f: files){
-            try {
-                Files.delete(f.toPath());
-                log.info("Files ts of live stream have been deleted");
-            } catch (IOException e) {
-                log.error("Files ts of live stream have not been deleted");
-                e.printStackTrace();
+        if(files != null) {
+            for (File f : files) {
+                try {
+                    Files.delete(f.toPath());
+                    log.info("Files ts of live stream have been deleted");
+                } catch (IOException e) {
+                    log.error("Files ts of live stream have not been deleted", e);
+                }
             }
         }
 
@@ -86,10 +79,7 @@ public class FFMpegCaptureStream extends AbstractFFMpegCapture {
             Files.delete(this.file.toPath());
             log.info("File m3u8 of live stream has been deleted");
         } catch (IOException e) {
-            log.error("File m3u8 of live stream has not been deleted");
-            e.printStackTrace();
+            log.error("File m3u8 of live stream has not been deleted", e);
         }
-
-
     }
 }
